Enable reuse task to optionally use non-exact matches
The Reuse task makes a call to server.GetUploadNum to identify
the prior and current uploads. This commit adds a configurable
flag to allow the Reuse task to optionally use non-exact
matching for the upload name. Default behavior will be identical
to prior status, but now the task can be configured to use the
non-exact matching if desired.

Signed-off-by: Steve Winslow <swinslow@gmail.com>