fix(tasks): Fixes #24 by checking server version
When setting up a BulkTextMatch action, the task will now check
the server version first. If it is at least 3.6.0, then the
action type's first letter will be capitalized ("Add" or "Remove")
else it will remain lowercase ("add" or "remove").

If the error from #24 is seen for versions between 3.3.0 and
3.6.0, then the minimum version check can be modified accordingly.

Signed-off-by: Steve Winslow <swinslow@gmail.com>